requirements.txt
appdirs>=1.4.3
Pillow>=5.2.0